# Conflicts:
#	src/engine/controls/cuecontrol.cpp

+++ src/library/dao/cuedao.cpp
@@ -43,8 +43,8 @@ CuePointer cueFromRow(const QSqlRecord& row) {
     int type = row.value(row.indexOf("type")).toInt();
     const auto position =
             mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(
-                    row.value(row.indexOf("position")).toInt());
-    int lengthFrames = row.value(row.indexOf("length")).toInt() / mixxx::kEngineChannelCount;
+                    row.value(row.indexOf("position")).toDouble());
+    double lengthFrames = row.value(row.indexOf("length")).toDouble() / mixxx::kEngineChannelCount;
     int hotcue = row.value(row.indexOf("hotcue")).toInt();
     QString label = labelFromQVariant(row.value(row.indexOf("label")));
     mixxx::RgbColor::optional_t color = mixxx::RgbColor::fromQVariant(row.value(row.indexOf("color")));
